requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pytest==7.4.0
starlette==0.27.0

# Day 2 RAG Dependencies
chromadb==0.4.15
sentence-transformers==5.1.0

# Day 3 
pytest-asyncio==0.21.1
coverage==7.3.2
python-multipart==0.0.20
huggingface_hub==0.35.0

# Advanced Features
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
torchaudio>=2.0.0,<3.0.0
ultralytics>=8.0.0,<9.0.0
opencv-python>=4.8.0,<5.0.0
Pillow>=9.5.0,<11.0.0
scikit-image>=0.21.0,<1.0.0
easyocr>=1.7.0,<2.0.0
networkx>=3.1,<4.0
matplotlib>=3.7.0,<4.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0